requirements-gemma.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.7
google-cloud-aiplatform==2.18.0
requests==2.31.0
